move appt specific date fns to appt helper; update tests
