+++ src/applications/travel-pay/tests/util/appointment-helpers.unit.spec.js
@@ -4,6 +4,8 @@ import {
   getPractionerName,
   getTimezoneByFacilityId,
   transformVAOSAppointment,
+  isPastAppt,
+  getDaysLeft,
 } from '../../util/appointment-helpers';
 
 const appointment = require('../fixtures/appointment.json');
@@ -34,6 +36,71 @@ describe('getTimezoneByFacilityId', () => {
   });
 });
 
+describe('isPastAppt', () => {
+  const appt = appointment.data.attributes;
+  const videoAppt = { ...appt, kind: 'telehealth' };
+
+  // The date in the appt is "2024-12-30T14:00:00Z"
+
+  afterEach(() => {
+    MockDate.reset();
+  });
+
+  it('returns true for appointment last month', () => {
+    MockDate.set('2025-01-30T15:00:00Z');
+    expect(isPastAppt(appt)).to.be.true;
+  });
+
+  it('returns false for appointment in the future', () => {
+    MockDate.set('2024-12-01T15:00:00Z');
+    expect(isPastAppt(appt)).to.be.false;
+  });
+
+  it('returns true for clinic appt 2 hours ago', () => {
+    MockDate.set('2024-12-30T16:00:00Z');
+    expect(isPastAppt(appt)).to.be.true;
+  });
+
+  it('returns false for clinic appt 30 minutes ago', () => {
+    MockDate.set('2024-12-30T14:30:00Z');
+    expect(isPastAppt(appt)).to.be.false;
+  });
+
+  it('returns true for video appt 5 hours ago', () => {
+    MockDate.set('2024-12-30T19:00:00Z');
+    expect(isPastAppt(videoAppt)).to.be.true;
+  });
+
+  it('returns false for video appt 2 hours ago', () => {
+    MockDate.set('2024-12-30T16:00:00Z');
+    expect(isPastAppt(videoAppt)).to.be.false;
+  });
+});
+
+describe('getDaysLeft', () => {
+  afterEach(() => {
+    MockDate.reset();
+  });
+
+  it('returns 10 for a date 20 days ago', () => {
+    MockDate.set('2024-06-25T15:00:00Z');
+    const actual = getDaysLeft('2024-06-05T14:00:00Z');
+    expect(actual).to.eq(10);
+  });
+
+  it('returns 30 for an appointment on the day filed', () => {
+    MockDate.set('2024-06-25T15:00:00Z');
+    const actual = getDaysLeft('2024-06-25T14:00:00Z');
+    expect(actual).to.eq(30);
+  });
+
+  it('returns 0 for a date more than 30 days ago', () => {
+    MockDate.set('2024-06-25T14:00:00Z');
+    const actual = getDaysLeft('2024-05-05T14:00:00Z');
+    expect(actual).to.eq(0);
+  });
+});
+
 describe('transformVAOSAppointment', () => {
   const appt = appointment.data.attributes;
 
@@ -48,6 +115,8 @@ describe('transformVAOSAppointment', () => {
 
     expect(transformedAppt).to.have.property('isPast');
     expect(transformedAppt.isPast).to.be.true;
+    expect(transformedAppt).to.have.property('daysSinceAppt');
+    expect(transformedAppt.daysSinceAppt).to.eq(16);
     expect(transformedAppt).to.have.property('isCC');
     expect(transformedAppt.isCC).to.be.false;
     expect(transformedAppt).to.have.property('isAtlas');
@@ -140,5 +209,6 @@ describe('transformVAOSAppointment', () => {
     const transformedAppt = transformVAOSAppointment(appt);
 
     expect(transformedAppt.isPast).to.be.false;
+    expect(transformedAppt.daysSinceAppt).to.be.null;
   });
 });
