+++ src/applications/travel-pay/tests/util/dates.unit.spec.js
@@ -1,13 +1,6 @@
 import MockDate from 'mockdate';
 import { expect } from 'chai';
-import {
-  getDateFilters,
-  formatDateTime,
-  getDaysLeft,
-  isPastAppt,
-} from '../../util/dates';
-
-const appointment = require('../fixtures/appointment.json');
+import { getDateFilters, formatDateTime } from '../../util/dates';
 
 function formatDateRange(dateRange) {
   const [startDay, startTime] = formatDateTime(dateRange.start);
@@ -210,71 +203,6 @@ describe('getDateFilters', () => {
     );
 
     expect(dateRanges.length).to.eq(6);
-  });
-});
-
-describe('getDaysLeft', () => {
-  afterEach(() => {
-    MockDate.reset();
-  });
-
-  it('returns 10 for a date 20 days ago', () => {
-    MockDate.set('2024-06-25T15:00:00Z');
-    const actual = getDaysLeft('2024-06-05T14:00:00Z');
-    expect(actual).to.eq(10);
-  });
-
-  it('returns 30 for an appointment on the day filed', () => {
-    MockDate.set('2024-06-25T15:00:00Z');
-    const actual = getDaysLeft('2024-06-25T14:00:00Z');
-    expect(actual).to.eq(30);
-  });
-
-  it('returns 0 for a date more than 30 days ago', () => {
-    MockDate.set('2024-06-25T14:00:00Z');
-    const actual = getDaysLeft('2024-05-05T14:00:00Z');
-    expect(actual).to.eq(0);
-  });
-});
-
-describe('isPastAppt', () => {
-  const appt = appointment.data.attributes;
-  const videoAppt = { ...appt, kind: 'telehealth' };
-
-  // The date in the appt is "2024-12-30T14:00:00Z"
-
-  afterEach(() => {
-    MockDate.reset();
-  });
-
-  it('returns true for appointment last month', () => {
-    MockDate.set('2025-01-30T15:00:00Z');
-    expect(isPastAppt(appt)).to.be.true;
-  });
-
-  it('returns false for appointment in the future', () => {
-    MockDate.set('2024-12-01T15:00:00Z');
-    expect(isPastAppt(appt)).to.be.false;
-  });
-
-  it('returns true for clinic appt 2 hours ago', () => {
-    MockDate.set('2024-12-30T16:00:00Z');
-    expect(isPastAppt(appt)).to.be.true;
-  });
-
-  it('returns false for clinic appt 30 minutes ago', () => {
-    MockDate.set('2024-12-30T14:30:00Z');
-    expect(isPastAppt(appt)).to.be.false;
-  });
-
-  it('returns true for video appt 5 hours ago', () => {
-    MockDate.set('2024-12-30T19:00:00Z');
-    expect(isPastAppt(videoAppt)).to.be.true;
-  });
-
-  it('returns false for video appt 2 hours ago', () => {
-    MockDate.set('2024-12-30T16:00:00Z');
-    expect(isPastAppt(videoAppt)).to.be.false;
   });
 });
