+++ src/applications/travel-pay/util/dates.js
@@ -1,12 +1,9 @@
 import {
-  addMinutes,
-  differenceInCalendarDays,
   endOfQuarter,
   endOfYear,
   format,
   getQuarter,
   getYear,
-  isBefore,
   startOfQuarter,
   startOfYear,
   subMonths,
@@ -15,7 +12,6 @@ import {
 } from 'date-fns';
 
 import { utcToZonedTime } from 'date-fns-tz';
-import { getTimezoneByFacilityId } from './appointment-helpers';
 
 export function formatDateTime(datetimeString, stripUTCIndicator = false) {
   const str = stripUTCIndicator
@@ -80,29 +76,4 @@ export function getDateFilters() {
   }
 
   return dateRanges;
-}
-
-export function getDaysLeft(datetimeString) {
-  const apptDate = new Date(datetimeString);
-  const daysSinceAppt = differenceInCalendarDays(new Date(), apptDate);
-
-  return daysSinceAppt > 30 ? 0 : 30 - daysSinceAppt;
-}
-
-export function isPastAppt(appointment) {
-  const isVideo = appointment.kind && appointment.kind === 'telehealth';
-  const threshold = isVideo ? 240 : 60;
-
-  const TZ = getTimezoneByFacilityId(appointment.locationId);
-
-  const startDate = TZ
-    ? new Date(appointment.start).toLocaleString('en-US', {
-        timeZone: TZ,
-      })
-    : new Date(appointment.start).toLocaleString();
-  const now = TZ
-    ? new Date().toLocaleString('en-US', { timeZone: TZ })
-    : new Date().toLocaleString();
-
-  return isBefore(addMinutes(new Date(startDate), threshold), new Date(now));
 }
